git-svn-id: file:///mnt/c/Users/svgroot/Downloads/svn/repos/Jumbo/branches/custom-rpc@862 2f0f0f96-4b5b-11dd-99f4-89bc9257463c

+++ NameServer/ReplicaPlacement.cs
@@ -30,7 +30,7 @@ namespace NameServerApplication
 
         public BlockAssignment AssignBlockToDataServers(IEnumerable<DataServerInfo> dataServers, BlockInfo block, string writerHostName, bool useLocalReplica)
         {
-            long freeSpaceThreshold = _configuration.NameServer.DataServerFreeSpaceThreshold;
+            long freeSpaceThreshold = (long)_configuration.NameServer.DataServerFreeSpaceThreshold;
             Guid blockId = block.BlockId;
             bool forceDifferentRack = false;
             var currentDataServers = (from server in dataServers
